requirements.txt
django
whitenoise
fabric
djangorestframework
markdown
django-filter